Fixed gulp-sourcemaps version
To days ago the version 1.7 of gulp-sourcemaps was released, but it is not working with our conifguration.
(There is no possibility to run any of the grunt tasks because of SyntaxError in gulp-sourcemaps/index.js).
I changed our package.json to use always 1.6 version, not 1.6 or higher as it was previously set.