OLMIS-1341: Fixed tokens being appended to templates
I've moved the templates to the openlmis top module, which
made most sense. Thanks to this tests started to fail - I figured
out the issue was with tokens being appended to .html template calls.
This a last ditch effort fix, so it might not be perfect.

If this won't fix the problem, I'm reverting the whole forsaken thing.