OLMIS-1348 Standardize action button for all forms
To standardize the UI, all forms should have at least one button to show action for the form. If the action button is conditional, remove the condition. If it can be disabled, remove that. Since the approval page form is different from the rest and it does not seem to work, remove it for now; it can be re-added once we know we need it.